refactor: [#164] move single-type dependency tests to dedicated modules
- Move 29 tests from core.rs to appropriate module files:
  - 3 tests → progress.rs (ProgressMessage tests)
  - 1 test → success.rs (SuccessMessage tests)
  - 1 test → error.rs (ErrorMessage tests)
  - 2 tests → result.rs (ResultMessage tests)
  - 1 test → warning.rs (WarningMessage tests)
  - 1 test → channel.rs (Channel enum tests)
  - 10 tests → steps.rs (StepsMessage tests)
  - 9 tests → info_block.rs (InfoBlockMessage tests)
  - 2 tests → composite.rs (CompositeSink tests)
  - 1 test → telemetry.rs (TelemetrySink tests)

- Remove all duplicate tests from core.rs (350+ lines reduction)
- Improve test organization by placing tests with their related types
- Maintain 100% test coverage with all 137 user_output tests passing
- Clean up module structure for better maintainability
