+++ src/presentation/user_output/core.rs
@@ -811,108 +811,6 @@ mod tests {
         use crate::presentation::user_output::test_support::TestUserOutput;
 
         #[test]
-        fn progress_message_should_format_with_theme() {
-            let theme = Theme::emoji();
-            let message = ProgressMessage {
-                text: "Test message".to_string(),
-            };
-
-            let formatted = message.format(&theme);
-
-            assert_eq!(formatted, "⏳ Test message\n");
-        }
-
-        #[test]
-        fn progress_message_should_require_normal_verbosity() {
-            let message = ProgressMessage {
-                text: "Test".to_string(),
-            };
-
-            assert_eq!(message.required_verbosity(), VerbosityLevel::Normal);
-        }
-
-        #[test]
-        fn progress_message_should_use_stderr_channel() {
-            let message = ProgressMessage {
-                text: "Test".to_string(),
-            };
-
-            assert_eq!(message.channel(), Channel::Stderr);
-        }
-
-        #[test]
-        fn success_message_should_format_with_theme() {
-            let theme = Theme::plain();
-            let message = SuccessMessage {
-                text: "Operation complete".to_string(),
-            };
-
-            let formatted = message.format(&theme);
-
-            assert_eq!(formatted, "[OK] Operation complete\n");
-        }
-
-        #[test]
-        fn error_message_should_always_be_shown() {
-            let message = ErrorMessage {
-                text: "Critical error".to_string(),
-            };
-
-            // Errors should require Quiet level (always shown)
-            assert_eq!(message.required_verbosity(), VerbosityLevel::Quiet);
-        }
-
-        #[test]
-        fn result_message_should_use_stdout_channel() {
-            let message = ResultMessage {
-                text: "Output data".to_string(),
-            };
-
-            assert_eq!(message.channel(), Channel::Stdout);
-        }
-
-        #[test]
-        fn result_message_should_not_include_symbols() {
-            let theme = Theme::emoji();
-            let message = ResultMessage {
-                text: "Plain output".to_string(),
-            };
-
-            let formatted = message.format(&theme);
-
-            // Result messages should not include theme symbols
-            assert!(!formatted.contains("⏳"));
-            assert!(!formatted.contains("✅"));
-            assert_eq!(formatted, "Plain output\n");
-        }
-
-        #[test]
-        fn steps_message_should_format_numbered_list() {
-            let theme = Theme::emoji();
-            let message = StepsMessage {
-                title: "Next steps:".to_string(),
-                items: vec!["First step".to_string(), "Second step".to_string()],
-            };
-
-            let formatted = message.format(&theme);
-
-            assert_eq!(formatted, "Next steps:\n1. First step\n2. Second step\n");
-        }
-
-        #[test]
-        fn warning_message_should_include_extra_space() {
-            let theme = Theme::emoji();
-            let message = WarningMessage {
-                text: "Warning text".to_string(),
-            };
-
-            let formatted = message.format(&theme);
-
-            // Warning messages include two spaces after the symbol
-            assert_eq!(formatted, "⚠️  Warning text\n");
-        }
-
-        #[test]
         fn user_output_write_should_respect_verbosity_filter() {
             let mut test_output = TestUserOutput::new(VerbosityLevel::Quiet);
 
@@ -947,13 +845,6 @@ mod tests {
 
             assert_eq!(test_output.stderr(), "⏳ Progress\n");
             assert_eq!(test_output.stdout(), "Result\n");
-        }
-
-        #[test]
-        fn channel_enum_should_support_equality() {
-            assert_eq!(Channel::Stdout, Channel::Stdout);
-            assert_eq!(Channel::Stderr, Channel::Stderr);
-            assert_ne!(Channel::Stdout, Channel::Stderr);
         }
 
         // Custom message type to demonstrate extensibility
@@ -1541,361 +1432,6 @@ mod tests {
     }
 
     // ============================================================================
-    // Builder Pattern Tests
-    // ============================================================================
-
-    mod builder_pattern {
-        use super::super::*;
-        use crate::presentation::user_output::formatters::JsonFormatter;
-        use crate::presentation::user_output::test_support::{self, TestUserOutput};
-
-        // ========================================================================
-        // StepsMessageBuilder Tests
-        // ========================================================================
-
-        #[test]
-        fn it_should_build_steps_with_fluent_api() {
-            let message = StepsMessage::builder("Title")
-                .add("Step 1")
-                .add("Step 2")
-                .add("Step 3")
-                .build();
-
-            assert_eq!(message.title, "Title");
-            assert_eq!(message.items, vec!["Step 1", "Step 2", "Step 3"]);
-        }
-
-        #[test]
-        fn it_should_create_simple_steps_directly() {
-            let message =
-                StepsMessage::new("Title", vec!["Step 1".to_string(), "Step 2".to_string()]);
-
-            assert_eq!(message.title, "Title");
-            assert_eq!(message.items, vec!["Step 1", "Step 2"]);
-        }
-
-        #[test]
-        fn it_should_build_empty_steps() {
-            let message = StepsMessage::builder("Title").build();
-
-            assert_eq!(message.title, "Title");
-            assert!(message.items.is_empty());
-        }
-
-        #[test]
-        fn it_should_build_single_step() {
-            let message = StepsMessage::builder("Title").add("Single step").build();
-
-            assert_eq!(message.title, "Title");
-            assert_eq!(message.items, vec!["Single step"]);
-        }
-
-        #[test]
-        fn it_should_accept_string_types_in_builder() {
-            let message = StepsMessage::builder("Title")
-                .add("String literal")
-                .add(String::from("Owned string"))
-                .add("Another literal".to_string())
-                .build();
-
-            assert_eq!(message.items.len(), 3);
-        }
-
-        #[test]
-        fn it_should_accept_string_types_in_constructor() {
-            let message =
-                StepsMessage::new("Title", vec!["Step 1".to_string(), String::from("Step 2")]);
-
-            assert_eq!(message.items.len(), 2);
-        }
-
-        #[test]
-        fn it_should_format_builder_messages_correctly() {
-            let theme = Theme::emoji();
-            let message = StepsMessage::builder("Next steps:")
-                .add("Configure")
-                .add("Deploy")
-                .build();
-
-            let formatted = message.format(&theme);
-            assert!(formatted.contains("Next steps:"));
-            assert!(formatted.contains("1. Configure"));
-            assert!(formatted.contains("2. Deploy"));
-        }
-
-        #[test]
-        fn it_should_integrate_builder_with_user_output() {
-            let mut test_output = TestUserOutput::new(VerbosityLevel::Normal);
-
-            let message = StepsMessage::builder("Next steps:")
-                .add("Edit config")
-                .add("Run tests")
-                .build();
-
-            test_output.output.write(&message);
-
-            let stderr = test_output.stderr();
-            assert!(stderr.contains("Next steps:"));
-            assert!(stderr.contains("1. Edit config"));
-            assert!(stderr.contains("2. Run tests"));
-        }
-
-        // ========================================================================
-        // InfoBlockMessageBuilder Tests
-        // ========================================================================
-
-        #[test]
-        fn it_should_build_info_block_with_fluent_api() {
-            let message = InfoBlockMessage::builder("Environment")
-                .add_line("Name: production")
-                .add_line("Status: active")
-                .build();
-
-            assert_eq!(message.title, "Environment");
-            assert_eq!(message.lines, vec!["Name: production", "Status: active"]);
-        }
-
-        #[test]
-        fn it_should_create_simple_info_block_directly() {
-            let message = InfoBlockMessage::new(
-                "Environment",
-                vec!["Name: production".to_string(), "Status: active".to_string()],
-            );
-
-            assert_eq!(message.title, "Environment");
-            assert_eq!(message.lines, vec!["Name: production", "Status: active"]);
-        }
-
-        #[test]
-        fn it_should_build_empty_info_block() {
-            let message = InfoBlockMessage::builder("Title").build();
-
-            assert_eq!(message.title, "Title");
-            assert!(message.lines.is_empty());
-        }
-
-        #[test]
-        fn it_should_build_single_line_info_block() {
-            let message = InfoBlockMessage::builder("Title")
-                .add_line("Single line")
-                .build();
-
-            assert_eq!(message.title, "Title");
-            assert_eq!(message.lines, vec!["Single line"]);
-        }
-
-        #[test]
-        fn it_should_accept_string_types_in_info_block_builder() {
-            let message = InfoBlockMessage::builder("Title")
-                .add_line("String literal")
-                .add_line(String::from("Owned string"))
-                .add_line("Another literal".to_string())
-                .build();
-
-            assert_eq!(message.lines.len(), 3);
-        }
-
-        #[test]
-        fn it_should_accept_string_types_in_info_block_constructor() {
-            let message =
-                InfoBlockMessage::new("Title", vec!["Line 1".to_string(), String::from("Line 2")]);
-
-            assert_eq!(message.lines.len(), 2);
-        }
-
-        #[test]
-        fn it_should_format_info_block_messages_correctly() {
-            let theme = Theme::emoji();
-            let message = InfoBlockMessage::builder("Environment")
-                .add_line("Name: production")
-                .add_line("Status: active")
-                .build();
-
-            let formatted = message.format(&theme);
-            assert!(formatted.contains("Environment"));
-            assert!(formatted.contains("Name: production"));
-            assert!(formatted.contains("Status: active"));
-        }
-
-        #[test]
-        fn it_should_integrate_info_block_builder_with_user_output() {
-            let mut test_output = TestUserOutput::new(VerbosityLevel::Normal);
-
-            let message = InfoBlockMessage::builder("Configuration")
-                .add_line("  - username: torrust")
-                .add_line("  - port: 22")
-                .build();
-
-            test_output.output.write(&message);
-
-            let stderr = test_output.stderr();
-            assert!(stderr.contains("Configuration"));
-            assert!(stderr.contains("  - username: torrust"));
-            assert!(stderr.contains("  - port: 22"));
-        }
-
-        #[test]
-        fn it_should_show_info_block_message_has_correct_properties() {
-            let message = InfoBlockMessage::new("Title", vec!["Line 1".to_string()]);
-
-            assert_eq!(message.required_verbosity(), VerbosityLevel::Normal);
-            assert_eq!(message.channel(), Channel::Stderr);
-            assert_eq!(message.type_name(), "InfoBlockMessage");
-        }
-
-        #[test]
-        fn it_should_respect_verbosity_for_info_block_messages() {
-            let mut test_output = TestUserOutput::new(VerbosityLevel::Quiet);
-
-            let message = InfoBlockMessage::builder("Info").add_line("Line 1").build();
-
-            test_output.output.write(&message);
-
-            // Should not appear at Quiet level
-            assert_eq!(test_output.stderr(), "");
-        }
-
-        #[test]
-        fn it_should_show_info_block_at_normal_level() {
-            let mut test_output = TestUserOutput::new(VerbosityLevel::Normal);
-
-            let message = InfoBlockMessage::builder("Info").add_line("Line 1").build();
-
-            test_output.output.write(&message);
-
-            // Should appear at Normal level
-            assert!(!test_output.stderr().is_empty());
-            assert!(test_output.stderr().contains("Info"));
-        }
-
-        // ========================================================================
-        // Backward Compatibility Tests
-        // ========================================================================
-
-        #[test]
-        fn it_should_maintain_backward_compatibility_for_steps() {
-            // Old way: direct construction
-            let old_message = StepsMessage {
-                title: "Steps".to_string(),
-                items: vec!["Step 1".to_string()],
-            };
-
-            // New way: constructor
-            let new_message = StepsMessage::new("Steps", vec!["Step 1".to_string()]);
-
-            // Should produce identical results
-            assert_eq!(old_message.title, new_message.title);
-            assert_eq!(old_message.items, new_message.items);
-        }
-
-        #[test]
-        fn it_should_maintain_backward_compatibility_for_info_blocks() {
-            // Old way: UserOutput::info_block helper
-            let mut test_output = TestUserOutput::new(VerbosityLevel::Normal);
-            test_output
-                .output
-                .info_block("Title", &["Line 1", "Line 2"]);
-            let old_output = test_output.stderr();
-
-            // New way: Direct message construction
-            let mut test_output = TestUserOutput::new(VerbosityLevel::Normal);
-            let message =
-                InfoBlockMessage::new("Title", vec!["Line 1".to_string(), "Line 2".to_string()]);
-            test_output.output.write(&message);
-            let new_output = test_output.stderr();
-
-            // Should produce identical output
-            assert_eq!(old_output, new_output);
-        }
-
-        // ========================================================================
-        // Integration Tests
-        // ========================================================================
-
-        #[test]
-        fn it_should_work_with_json_formatter() {
-            use parking_lot::Mutex;
-            use std::sync::Arc;
-
-            let stdout_buffer = Arc::new(Mutex::new(Vec::new()));
-            let stderr_buffer = Arc::new(Mutex::new(Vec::new()));
-            let formatter = Box::new(JsonFormatter);
-
-            let mut output = UserOutput {
-                theme: Theme::emoji(),
-                verbosity_filter: VerbosityFilter::new(VerbosityLevel::Normal),
-                sink: Box::new(StandardSink::new(
-                    Box::new(test_support::TestWriter::new(Arc::clone(&stdout_buffer))),
-                    Box::new(test_support::TestWriter::new(Arc::clone(&stderr_buffer))),
-                )),
-                formatter_override: Some(formatter),
-            };
-
-            let message = StepsMessage::builder("Steps").add("Step 1").build();
-            output.write(&message);
-
-            let stderr = String::from_utf8(stderr_buffer.lock().clone()).unwrap();
-            let json: serde_json::Value = serde_json::from_str(stderr.trim()).unwrap();
-
-            assert_eq!(json["type"], "StepsMessage");
-        }
-
-        #[test]
-        fn it_should_work_with_info_block_json_formatter() {
-            use parking_lot::Mutex;
-            use std::sync::Arc;
-
-            let stdout_buffer = Arc::new(Mutex::new(Vec::new()));
-            let stderr_buffer = Arc::new(Mutex::new(Vec::new()));
-            let formatter = Box::new(JsonFormatter);
-
-            let mut output = UserOutput {
-                theme: Theme::emoji(),
-                verbosity_filter: VerbosityFilter::new(VerbosityLevel::Normal),
-                sink: Box::new(StandardSink::new(
-                    Box::new(test_support::TestWriter::new(Arc::clone(&stdout_buffer))),
-                    Box::new(test_support::TestWriter::new(Arc::clone(&stderr_buffer))),
-                )),
-                formatter_override: Some(formatter),
-            };
-
-            let message = InfoBlockMessage::builder("Info").add_line("Line 1").build();
-            output.write(&message);
-
-            let stderr = String::from_utf8(stderr_buffer.lock().clone()).unwrap();
-            let json: serde_json::Value = serde_json::from_str(stderr.trim()).unwrap();
-
-            assert_eq!(json["type"], "InfoBlockMessage");
-        }
-
-        #[test]
-        fn it_should_handle_many_items_in_builder() {
-            let mut builder = StepsMessage::builder("Many steps");
-            for i in 1..=100 {
-                builder = builder.add(format!("Step {i}"));
-            }
-            let message = builder.build();
-
-            assert_eq!(message.items.len(), 100);
-            assert_eq!(message.items[0], "Step 1");
-            assert_eq!(message.items[99], "Step 100");
-        }
-
-        #[test]
-        fn it_should_handle_many_lines_in_info_block_builder() {
-            let mut builder = InfoBlockMessage::builder("Many lines");
-            for i in 1..=100 {
-                builder = builder.add_line(format!("Line {i}"));
-            }
-            let message = builder.build();
-
-            assert_eq!(message.lines.len(), 100);
-            assert_eq!(message.lines[0], "Line 1");
-            assert_eq!(message.lines[99], "Line 100");
-        }
-    }
-
-    // ============================================================================
     // OutputSink Tests
     // ============================================================================
 
@@ -2015,44 +1551,6 @@ mod tests {
             assert_eq!(messages1.lock()[0], "⏳ Test\n");
             assert_eq!(messages2.lock()[0], "⏳ Test\n");
             assert_eq!(messages3.lock()[0], "⏳ Test\n");
-        }
-
-        #[test]
-        fn composite_sink_should_support_empty_sink_list() {
-            let mut composite = CompositeSink::new(vec![]);
-
-            let message = ProgressMessage {
-                text: "Test".to_string(),
-            };
-            let theme = Theme::emoji();
-            let formatted = message.format(&theme);
-
-            // Should not panic with empty sink list
-            composite.write_message(&message, &formatted);
-        }
-
-        #[test]
-        fn composite_sink_should_support_add_sink() {
-            let messages1 = Arc::new(Mutex::new(Vec::new()));
-            let messages2 = Arc::new(Mutex::new(Vec::new()));
-
-            let mut composite =
-                CompositeSink::new(vec![Box::new(MockSink::new(Arc::clone(&messages1)))]);
-
-            // Add another sink
-            composite.add_sink(Box::new(MockSink::new(Arc::clone(&messages2))));
-
-            let message = ProgressMessage {
-                text: "Test".to_string(),
-            };
-            let theme = Theme::emoji();
-            let formatted = message.format(&theme);
-
-            composite.write_message(&message, &formatted);
-
-            // Verify both sinks received the message
-            assert_eq!(messages1.lock().len(), 1);
-            assert_eq!(messages2.lock().len(), 1);
         }
 
         #[test]
@@ -2229,12 +1727,6 @@ mod tests {
         // ========================================================================
         // TelemetrySink Tests
         // ========================================================================
-
-        #[test]
-        fn telemetry_sink_should_create_with_endpoint() {
-            let sink = TelemetrySink::new("https://example.com".to_string());
-            assert_eq!(sink.endpoint(), "https://example.com");
-        }
 
         #[test]
         fn telemetry_sink_should_log_messages() {
