+++ src/presentation/user_output/messages/steps.rs
@@ -201,3 +201,119 @@ impl StepsMessageBuilder {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn steps_message_should_format_numbered_list() {
+        let theme = Theme::emoji();
+        let message = StepsMessage {
+            title: "Next steps:".to_string(),
+            items: vec!["First step".to_string(), "Second step".to_string()],
+        };
+
+        let formatted = message.format(&theme);
+
+        assert_eq!(formatted, "Next steps:\n1. First step\n2. Second step\n");
+    }
+
+    #[test]
+    fn it_should_build_steps_with_fluent_api() {
+        let message = StepsMessage::builder("Title")
+            .add("Step 1")
+            .add("Step 2")
+            .add("Step 3")
+            .build();
+
+        assert_eq!(message.title, "Title");
+        assert_eq!(message.items, vec!["Step 1", "Step 2", "Step 3"]);
+    }
+
+    #[test]
+    fn it_should_create_simple_steps_directly() {
+        let message = StepsMessage::new("Title", vec!["Step 1".to_string(), "Step 2".to_string()]);
+
+        assert_eq!(message.title, "Title");
+        assert_eq!(message.items, vec!["Step 1", "Step 2"]);
+    }
+
+    #[test]
+    fn it_should_build_empty_steps() {
+        let message = StepsMessage::builder("Title").build();
+
+        assert_eq!(message.title, "Title");
+        assert!(message.items.is_empty());
+    }
+
+    #[test]
+    fn it_should_build_single_step() {
+        let message = StepsMessage::builder("Title").add("Single step").build();
+
+        assert_eq!(message.title, "Title");
+        assert_eq!(message.items, vec!["Single step"]);
+    }
+
+    #[test]
+    fn it_should_accept_string_types_in_builder() {
+        let message = StepsMessage::builder("Title")
+            .add("String literal")
+            .add(String::from("Owned string"))
+            .add("Another literal".to_string())
+            .build();
+
+        assert_eq!(message.items.len(), 3);
+    }
+
+    #[test]
+    fn it_should_accept_string_types_in_constructor() {
+        let message =
+            StepsMessage::new("Title", vec!["Step 1".to_string(), String::from("Step 2")]);
+
+        assert_eq!(message.items.len(), 2);
+    }
+
+    #[test]
+    fn it_should_format_builder_messages_correctly() {
+        let theme = Theme::emoji();
+        let message = StepsMessage::builder("Next steps:")
+            .add("Configure")
+            .add("Deploy")
+            .build();
+
+        let formatted = message.format(&theme);
+        assert!(formatted.contains("Next steps:"));
+        assert!(formatted.contains("1. Configure"));
+        assert!(formatted.contains("2. Deploy"));
+    }
+
+    #[test]
+    fn it_should_maintain_backward_compatibility_for_steps() {
+        // Old way: direct construction
+        let old_message = StepsMessage {
+            title: "Steps".to_string(),
+            items: vec!["Step 1".to_string()],
+        };
+
+        // New way: constructor
+        let new_message = StepsMessage::new("Steps", vec!["Step 1".to_string()]);
+
+        // Should produce identical results
+        assert_eq!(old_message.title, new_message.title);
+        assert_eq!(old_message.items, new_message.items);
+    }
+
+    #[test]
+    fn it_should_handle_many_items_in_builder() {
+        let mut builder = StepsMessage::builder("Many steps");
+        for i in 1..=100 {
+            builder = builder.add(format!("Step {i}"));
+        }
+        let message = builder.build();
+
+        assert_eq!(message.items.len(), 100);
+        assert_eq!(message.items[0], "Step 1");
+        assert_eq!(message.items[99], "Step 100");
+    }
+}
