--- test-refactor-tracking.md
@@ -1,61 +1,0 @@
-# Test Refactor Tracking
-
-## Objective
-
-Move unit tests from src/presentation/user_output/core.rs to their respective struct modules, keep integration tests in core.rs.
-
-## Test Analysis Status
-
-- [x] `theme` module tests (7 tests) - COMPLETED - REMOVED (duplicates)
-- [x] `type_safe_wrappers` module tests (8 tests) - COMPLETED - SPLIT (6 moved, 2 kept)
-- [x] `verbosity_filter` module tests (18 tests) - COMPLETED - SPLIT (16 moved, 2 kept)
-- [x] `parameterized_tests` module tests (4 tests) - ANALYZED - INTEGRATION TESTS - KEPT
-- [x] Individual UserOutput tests (6 tests) - ANALYZED - MIXED - 4 VerbosityLevel unit tests moved, 2 integration tests kept
-- [x] `output_message_trait` module tests (15 tests) - ANALYZED - INTEGRATION TESTS - KEPT
-- [x] `user_output_with_themes` module tests (4 tests) - ANALYZED - INTEGRATION TESTS - KEPT
-- [x] `formatter_override` module tests (~8 tests) - ANALYZED - INTEGRATION TESTS - KEPT
-- [x] `buffering` module tests (5 tests) - ANALYZED - INTEGRATION TESTS - KEPT
-- [x] `builder_pattern` module tests (~20 tests) - ANALYZED - MIXED/INTEGRATION - KEPT
-- [x] `output_sink` module tests (~15 tests) - ANALYZED - MIXED/INTEGRATION - KEPT
-
-## Analysis Key
-
-- ✅ Unit Test (single struct) - Move to module
-- 🔗 Integration Test (multiple structs) - Keep in core.rs
-- 📝 Completed
-
-## SYSTEMATIC REVIEW COMPLETE ✅
-
-### Summary of Actions Taken
-
-1. **Theme Tests (7 tests)**: REMOVED from core.rs - All were duplicates of existing tests in theme.rs
-2. **Type-Safe Wrappers (8 tests)**: SPLIT - 6 unit tests moved to writers.rs, 2 integration tests kept in core.rs
-3. **VerbosityFilter (18 tests)**: SPLIT - 16 unit tests moved to verbosity.rs, 2 integration tests kept in core.rs
-
-### Modules Analyzed and Kept as Integration Tests
-
-1. **Parameterized Tests**: Tests UserOutput with various theme/verbosity combinations using rstest
-1. **Individual UserOutput Tests**: Tests UserOutput methods with complete component stack (2 integration tests) and VerbosityLevel trait tests (4 unit tests moved to verbosity.rs)
-1. **OutputMessage Trait Tests**: Tests trait implementations across multiple message types
-1. **UserOutput with Themes Tests**: Tests UserOutput + Theme integration and formatting
-1. **Formatter Override Tests**: Tests UserOutput + JsonFormatter + FormatterOverride integration
-1. **Buffering Tests**: Tests UserOutput + buffer management integration with flush operations
-1. **Builder Pattern Tests**: Tests message builders with UserOutput integration and JSON formatting
-1. **Output Sink Tests**: Tests various sink implementations with UserOutput integration
-
-### Final Statistics
-
-- **Total test modules analyzed**: 11
-- **Unit tests moved**: 26 (6 type-safe wrappers + 16 verbosity filter + 4 verbosity level)
-- **Duplicate tests removed**: 7 (theme tests)
-- **Integration test modules preserved**: 8 modules containing ~65+ integration tests
-- **Core principle applied**: Unit tests (single struct) moved to struct modules, integration tests (multiple structs) kept in core.rs
-
-### Validation
-
-✅ All moved tests pass in their target modules
-✅ All integration tests continue to pass in core.rs  
-✅ No functionality lost during refactoring
-✅ Clear separation between unit and integration concerns achieved
-
-The systematic review is complete. The refactoring successfully achieved the goal of cleaning up core.rs by moving pure unit tests to appropriate modules while preserving the valuable integration tests that verify component interactions.
